add jupyter notebook about supervised classification using TFIDF
This document describes and implements a prototype to classify job ads
using the TF.IDF technique to create a vector space of words. In this
space the distance between classes can be measured and new documents
are classified by mapping their bag-of-words representation into a
vector in the same space with TF.IDF and computing the cosine or
jaccard distance to the nearest classes.